Stop creating and loading optimizer in default create_nerf function
To be able to load checkpoints for sampling network
optimizer must be created AFTER sampling network.
We need to pass grad_vars to optimizer so it matches
saved state.
Note that this function will not work on its own,
without running it from SamplingTrainer, where
optimizer is created there:
https://github.com/MarcinKadziolka/nerf-sampling/blob/99fe9bd114f8774dd2b516428622803bc73ba1ce/nerf_sampling/trainers/sampling_trainer.py#L102-L105

Actually without loading this approach worked as previously
optimizer was overwritten in SamplingTrainer function.
But trying to load optim before ST is created
resulted in error.