Start new loss function that focuses on maximum density instead of the mean
Upon examining histograms it seems that in original nerf despite that
most values of density are 0 or negative, there always are some samples
that place themselves within 700-1000 range of density, where the overall
mean density is around 11.

Using mean of density as new loss function indeed improves our mean
to higher values, but it does that by mostly reducing the number of
negative samples and we end up with all samples of density somewhere between
50-100. Goal is achieved as we surpass original nerf, but PSNR isn't great.

Therefore this approach will focus on maximizing max values to match
original nerf results but in less samples.